Renaming test classes
