+++ src/SonarQube.Client.RequestGenerator/Program.cs
@@ -1,4 +1,24 @@
-﻿using System;
+﻿/*
+ * SonarQube Client
+ * Copyright (C) 2016-2018 SonarSource SA
+ * mailto:info AT sonarsource DOT com
+ *
+ * This program is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU Lesser General Public
+ * License as published by the Free Software Foundation; either
+ * version 3 of the License, or (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+ * Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with this program; if not, write to the Free Software Foundation,
+ * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
+ */
+
+using System;
 using System.Net.Http;
 using System.Threading;
 using SonarQube.Client.Api.Requests;
@@ -6,20 +26,19 @@ using SonarQube.Client.Services;
 
 namespace SonarQube.Client.RequestGenerator
 {
-    class Program
+    public static class Program
     {
-        static void Main(string[] args)
+        private static void Main(string[] args)
         {
-            var requestFactory = new RequestFactory();
-            DefaultConfiguration.Configure(requestFactory);
-
             var cts = new CancellationTokenSource();
 
-            var service = new SonarQubeService(new HttpClientHandler(), requestFactory);
+            var sonarQubeService = new SonarQubeService(
+                new HttpClientHandler(),
+                DefaultConfiguration.Configure(new RequestFactory()));
 
             var parser = new ArgsParser(args);
 
-            var runner = new ServiceRunner(service)
+            var runner = new ServiceRunner(sonarQubeService)
             {
                 OutputPath = parser.NextArg(),
                 SonarQubeUrl = new Uri(parser.NextArg()),
